% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser.R
\name{browser_hook_sync_debugger}
\alias{browser_hook_sync_debugger}
\title{Generate a Debug Adapter Browser Hook}
\usage{
browser_hook_sync_debugger(debuggee)
}
\arguments{
\item{connection}{A connection to a debug adapter. This connection is used
to send information about the active debug state.}
}
\description{
Generate a Debug Adapter Browser Hook
}
