% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{dap}
\alias{dap}
\alias{run}
\alias{dap_attach}
\alias{dap_launch}
\alias{dap_launch_bg}
\title{Debug Adapter}
\usage{
run(
  host = getOption("debugadapter.host", "localhost"),
  port = getOption("debugadapter.port", 18721),
  ...,
  wait_for_client = FALSE
)

dap_attach(
  host = getOption("debugadapter.host", "localhost"),
  port = getOption("debugadapter.port", 18721),
  ...,
  process = NULL
)

dap_launch(...)

dap_launch_bg(..., host)
}
\arguments{
\item{host}{The host to use for connections}

\item{port}{The port to use for connections}

\item{...}{Arguments used to initialize an \link{adapter}}

\item{poll}{The polling frequency in seconds}

\item{timeout}{A timeout time in seconds to apply to all connections}
}
\description{
Starts a tcp server that listens for an arbitrary number of clients, and
exactly one R environment in which to perform debugging operations. When
serving as debugger in attach mode, the R environment itself may also
operate as its own debug client.
}
