% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{dapsink}
\alias{dapsink}
\title{Collection of utilities for managing redirection of default debugger output}
\format{
An object of class \code{environment} of length 6.
}
\usage{
dapsink
}
\arguments{
\item{set}{expects either "open" (to redirect stdout) or "close"}
}
\description{
Collection of utilities for managing redirection of default debugger output
}
\keyword{datasets}
