% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter.R
\name{debug_adapter}
\alias{debug_adapter}
\alias{$.debug_adapter}
\alias{$<-.debug_adapter}
\alias{next_id}
\alias{supports}
\alias{format.debug_adapter}
\title{Debug Adapter}
\usage{
debug_adapter(x)

\method{$}{debug_adapter}(x, name)

\method{$}{debug_adapter}(x, name) <- value

next_id(adapter, n = 1)

supports(x, capability)

\method{format}{debug_adapter}(x)
}
\arguments{
\item{x}{A connection object used to communicate with the debug client}
}
\description{
A simple class which stores state for a debug adapter. Operationally, this is
an environment for in-place mutation of the adapter state.
}
\section{Methods (by generic)}{
\itemize{
\item \code{$}: Accessors for the debug adapter

\item \code{`$`(debug_adapter) <- value}: Assignment to stateful components of the debug adapter

\item \code{format(debug_adapter)}: Convert a debug adapter to a string representation

}}
\section{Functions}{
\itemize{
\item \code{next_id()}: Generate a unique id number for breakpoints

\item \code{supports()}: Test whether adapter supports a capability

}}
\keyword{internal}
