% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debuggers.R
\name{debugger-interfaces}
\alias{debugger-interfaces}
\alias{debug_in_client_terminal}
\alias{debug_stdout_relay}
\alias{debug_here}
\title{Debugger Interfaces}
\usage{
debug_in_client_terminal()

debug_stdout_relay()

debug_here()
}
\description{
These classes are stubs used for dispatching on choice of debugger
interfaces. How the adapter interacts with clients can vary based on the
client that is used.
}
\section{Functions}{
\itemize{
\item \code{debug_in_client_terminal()}: Debugging commands are sent to a terminal managed by the client.

\item \code{debug_stdout_relay()}: Debugging is handled by emitting R code over stdout for another prcoess to
pick up and execute to handle debugging. By default, this is used when the
debugger is launched in an interactive session, launching the server in the
background, which then relays debug callbacks to the parent process.

\item \code{debug_here()}: Handle debugging calls locally ('here').

}}
