% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_source_object}
\alias{find_source_object}
\title{Find Source of Object}
\usage{
find_source_object(expr, envir)
}
\arguments{
\item{expr}{In most uses, an expression. Dispatch is used to recursively
discover source objects after evaluation, so after recursing, this might
be an R object.}

\item{envir}{An environment in which to evaluate the expression}
}
\description{
When debugging, many objects are modified with injected traces. We don't
care about these because they only exist in our R Session. What we want
is the original object whose source may be opened in a client editor.
}
