% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_variable}
\alias{format_variable}
\title{Format a variable for return as part of a Variable.value type}
\usage{
format_variable(x)
}
\arguments{
\item{x}{Any R object}
}
\description{
When sending variables in scope back to the client, we format them for
display.
}
