% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints.R
\name{get_breakpoint_locations}
\alias{get_breakpoint_locations}
\title{Find Source Locations based Breakpoint Location}
\usage{
get_breakpoint_locations(path, line)
}
\arguments{
\item{path}{A file path of the source code to trace}

\item{line}{The line number in the source file to trace}
}
\description{
A single breakpoint may be best served by tracing multiple source code
locations. For example, functions loaded from a package are traced
separately from the namespaced function calls, resulting in multiple
traced locations based on a single source and line number.
}
