% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{has_browser_hook}
\alias{has_browser_hook}
\title{Tests whether the current version of R has recent changes to the browser}
\usage{
has_browser_hook()
}
\value{
\code{logical[1L]} indicating whether R is of a satisfactory version
}
\description{
Tests whether the current version of R has recent changes to the browser
}
\note{
Specifically, tests whether argument \code{ignoreHook} is accepted by
\code{\link[=browser]{browser()}}. If this feature is reverted, this function will need to
be updated to respond to changes in devel.
}
