% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{insert_sync_callback}
\alias{insert_sync_callback}
\title{Insert a Task Callback to Synchronize Session}
\usage{
insert_sync_callback(debuggee, process)
}
\arguments{
\item{debuggee}{A debug adapter client, and specifically a debuggee - a
privileged client that is used as the source of debug information.}

\item{process}{Optionally, a process handle so that we can query the
background process running the debug adapter and surface any errors that
it may have encountered. Not strictly necessary, but very helpful.}
}
\description{
The debug adapter is constantly talking to editors in the background, but
it can only interject in our R session periodically to query our session
for new information.
}
\details{
Top-level task callbacks are one of few ways that we can periodically add
additional, recurring code to a typical R session.
}
