% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logging.R
\docType{data}
\name{logging}
\alias{logging}
\alias{loglevel}
\alias{TRACE}
\alias{DEBUG}
\alias{INFO}
\alias{log}
\alias{echo}
\title{Emit Logging Messages}
\format{
An object of class \code{list} of length 3.
}
\usage{
loglevel

TRACE(...)

DEBUG(...)

INFO(...)

log(level, ..., verbose = getOption("debugadapter.log", FALSE))

echo(level, msg, verbose = getOption("debugadapter.log", FALSE))
}
\arguments{
\item{...}{Any number of arguments to comprise the logging message. Atomics
are printed as is, and anything else is formatted before printing.}

\item{level}{The logging level that the message should be emitted with}

\item{verbose}{A verbosity level. When lower than the logging level, the
logging message is ignored.}
}
\description{
Emit Logging Messages

Echo Logging Messages
}
\keyword{datasets}
