% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{package-file-helpers}
\alias{package-file-helpers}
\alias{find_package_root}
\title{Package source file helpers}
\usage{
find_package_root(path = ".")
}
\arguments{
\item{path}{A path within a package source or install directory}
}
\description{
Discover specific package related file paths
}
\keyword{internal}
