% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{protocol-messages}
\alias{protocol-messages}
\alias{request}
\alias{response}
\alias{response_add_fields}
\alias{is_response}
\alias{event}
\title{Protocol Messages}
\usage{
request(command, arguments)

response(to, ...)

response_add_fields(to, body, ...)

is_response(x)

event(event, ...)
}
\arguments{
\item{command, event}{\code{string} relevant type information. Varies by
message type.}

\item{arguments, body}{\code{list} data to attach to the message. Varies by
message type.}

\item{to}{\code{request} When issueing a response, the first argument is the
request that is being responded to, used as the basis for communication.}

\item{x}{\code{list} an object}
}
\value{
A \code{list} matching the respective message structure
}
\description{
Constructing and working with protocol messages
}
\details{
The debug adapter protocol relies on a few broad classes of messages
including requests, responses, events and reverse requests. Each of these
has a rigid structure that is adapted for the individual behavior of each
message type.
}
