% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_session.R
\name{shadow}
\alias{shadow}
\title{Shadow browser}
\usage{
shadow(expr)
}
\arguments{
\item{expr}{a \dQuote{condition}.  By default, and whenever not false
    after being coerced to \code{\link[base]{logical}}, the
    debugger will be invoked, otherwise control is returned directly.}
}
\description{
Run \code{browser()} in a forked process, presenting a debugger in a familiar
\code{browser()} interface, while allowing additional metadata to be captured to
feed an IDE using the debug adapter protocol.
}
\examples{
x <- function(...) { y(...) }
y <- function(...) {
  1
  2
  3
  print("hello, world!")
}

x()

}
