% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_session.R
\name{shadow_browser_pipes}
\alias{shadow_browser_pipes}
\title{Create a pair of pipe pairs for bi-directional communication}
\usage{
shadow_browser_pipes()
}
\description{
Create a pair of pipe pairs for bi-directional communication
}
