% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter.R
\name{start_adapter}
\alias{start_adapter}
\alias{start_adapter_in_background}
\title{Debug Adapter}
\usage{
start_adapter(...)

start_adapter_in_background(...)
}
\arguments{
\item{...}{Arguments used to initialize an \link{adapter}}

\item{port}{The port to listen on for new connections}

\item{poll}{The polling frequency in seconds}

\item{timeout}{A timeout time in seconds to apply to all connections}
}
\description{
Starts a tcp server that listens for an arbitrary number of clients, and
exactly one R environment in which to perform debugging operations. When
serving as debugger in attach mode, the R environment itself may also
operate as its own debug client.
}
