% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints.R
\name{verify_breakpoint}
\alias{verify_breakpoint}
\title{Verify DAP sourceBreakpoint and return a Breakpoint}
\usage{
verify_breakpoint(path, line, id = NULL)
}
\description{
Verifying a breakpoint provided by the client means searching for the line of
code within the source file and verifying that a breakpoint can be placed
there.
}
\note{
The DAP sourceBreakpoint can provide a column, though this is currently
unimplemented.

(\url{https://microsoft.github.io/debug-adapter-protocol/specification#Requests_SetBreakpoints})
(\url{https://microsoft.github.io/debug-adapter-protocol/specification#Types_SourceBreakpoint})
(\url{https://microsoft.github.io/debug-adapter-protocol/specification#Types_Breakpoint})
}
