% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugger_foreground.R, R/debuggers.R
\name{debug_in_foreground}
\alias{debug_in_foreground}
\alias{debugger-interfaces}
\alias{debug_in_client_terminal}
\title{Debugger Interfaces}
\usage{
debug_in_foreground(con)

debug_in_client_terminal(con)
}
\description{
These classes are stubs used for dispatching on choice of debugger
interfaces. How the adapter interacts with clients can vary based on the
client that is used.
}
\section{Functions}{
\itemize{
\item \code{debug_in_foreground()}: Handle debugging calls locally ('here').

\item \code{debug_in_client_terminal()}: Debugging commands are sent to a terminal managed by the client.

}}
