% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{messages}
\alias{messages}
\alias{read_message}
\alias{read_message.client}
\alias{read_message.default}
\alias{write_message}
\alias{write_message.client}
\alias{write_message.default}
\alias{parse_message_body}
\alias{format_message_content}
\title{Protocol Message Reading and Parsing}
\usage{
read_message(x, ...)

\method{read_message}{client}(x, ...)

\method{read_message}{default}(x, ..., name = NULL, timeout = NULL, poll = timeout/100)

write_message(x, content = list(), ...)

\method{write_message}{client}(x, content = list(), ...)

\method{write_message}{default}(x, content = list(), name = NULL)

parse_message_body(content)

format_message_content(content)
}
\arguments{
\item{x}{A connection (or similar) to read from or write to}

\item{name}{An optional string to use for the connection for debug logs}

\item{timeout}{A maximum time in seconds to wait to receive a new message}

\item{content}{The message contents to write}
}
\description{
Protocol Message Reading and Parsing
}
