% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{handle}
\alias{handle}
\alias{handle.request.initialize}
\alias{handle.request.attach}
\alias{handle.request.setExceptionBreakpoints}
\alias{handle.request.setBreakpoints}
\alias{handle.request.disconnect}
\alias{handle.reverse_request}
\alias{handle.response}
\title{Dispatch for protocol messaging}
\usage{
handle(x, ...)

\method{handle}{request.initialize}(x, ..., adapter)

\method{handle}{request.attach}(x, ..., adapter)

\method{handle}{request.setExceptionBreakpoints}(x, ..., adapter)

\method{handle}{request.setBreakpoints}(x, ..., adapter)

\method{handle}{request.disconnect}(x, ..., adapter)

\method{handle}{reverse_request}(x, ..., adapter)

\method{handle}{response}(x, ...)
}
\description{
Dispatch for protocol messaging
}
\section{Functions}{
\itemize{
\item \code{handle(request.initialize)}: Initiailze the debug adapter and store client information for this session.
(\url{https://microsoft.github.io/debug-adapter-protocol/specification#Requests_Initialize})

\item \code{handle(request.attach)}: Recieve attach request from client and respond to confirm.
(\url{https://microsoft.github.io/debug-adapter-protocol/specification#Requests_Attach})

\item \code{handle(request.setExceptionBreakpoints)}: Update internal adapter breakpoints listing and relay breakpoints to debug
session.
(\url{https://microsoft.github.io/debug-adapter-protocol/specification#Requests_setExceptionBreakpoints})

\item \code{handle(request.setBreakpoints)}: Update internal adapter breakpoints listing and relay breakpoints to debug
session.
(\url{https://microsoft.github.io/debug-adapter-protocol/specification#Requests_SetBreakpoints})

\item \code{handle(request.disconnect)}: Handle disconnect request to disconnect from the debuggee, end the debug
session and shut itself down.
(\url{https://microsoft.github.io/debug-adapter-protocol/specification#Requests_Disconnect})

\item \code{handle(reverse_request)}: Issue reverse request and await response.
(\url{https://microsoft.github.io/debug-adapter-protocol/specification#Reverse_Requests_RunInTerminal})

\item \code{handle(response)}: Recieve reverse request response.
(\url{https://microsoft.github.io/debug-adapter-protocol/specification#Reverse_Requests_RunInTerminal})

}}
