% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser.R
\name{shadow_browser}
\alias{shadow_browser}
\title{Shadow browser}
\usage{
shadow_browser(..., ui = ui_browser_prompt)
}
\description{
Run \code{browser()} in a forked process, presenting a debugger in a familiar
\code{browser()} interface, while allowing additional metadata to be captured to
feed an IDE using the debug adapter protocol.
}
\details{
Unfortunately, navigating history with arrows, multiline input and using
other terminal-accepted inputs for line navigation will not work.
}
\examples{
x <- function() { y() }
y <- function() {
  shadow_browser(skipCalls = 4)
  print("hello, world")
}

x()

}
